resolving pr comments - making utils
