+++ frontend-two/utils/date.ts
@@ -1,5 +1,21 @@
 import { DateTime, Duration } from "luxon";
 
+export const parseDate = (raw: string | null, fallback: DateTime): DateTime => {
+  if (!raw) return fallback;
+  const parsed = DateTime.fromISO(raw, { zone: "utc" });
+  return parsed.isValid ? parsed : fallback;
+};
+
+export const toIsoDateInput = (dateTime: DateTime): string =>
+  dateTime.toISODate() ?? DateTime.utc().toISODate()!;
+
+export const formatTransitTime = (
+  seconds: number | null | undefined,
+): string => {
+  if (!seconds || seconds <= 0) return "Unavailable";
+  return Duration.fromObject({ seconds }).toFormat("mm:ss");
+};
+
 export const formatMinuteSeconds = (minute: number) =>
   Duration.fromObject({ minute }).toFormat("m:ss");
